fix change cosigners function
